requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
email-validator

# Database
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# LLM / AI
google-genai
groq
tiktoken

# File Parsing
pypdf
python-docx
beautifulsoup4

# ML
# numpy<2.0.0
# scikit-learn

# HTTP Client
requests
httpx

# Utilities
python-dateutil
tenacity

# Testing
pytest
pytest-asyncio